Refactor BoatnetHauls component: move columns and data to props
