+++ libs/bn-common/src/components/BoatnetHauls.vue
@@ -20,7 +20,7 @@
         </template>
       </q-input>
     </div>
-    <q-table :data="data" :columns="columns" :selected.sync="selected"/>
+    <q-table :data="haulsData" :columns="haulsSettings.columns" :selected.sync="selected"/>
 
     <!-- <div class="q-mt-md">Selected: {{ JSON.stringify(selected) }}</div> -->
     <div class="row">
@@ -39,146 +39,12 @@
 import { Component, Prop, Vue } from 'vue-property-decorator';
 import { BoatnetHaulsSettings } from '../models/BoatnetHaulsSettings';
 
-import moment from 'moment';
-
-import { Point } from 'geojson';
-import {
-  WcgopHaul,
-  WcgopHaulTypeName,
-  LocationEvent,
-  Vessel
-} from '@boatnet/bn-models';
-
 @Component
 export default class BoatnetHauls extends Vue {
-  @Prop() public programInfo: BoatnetHaulsSettings | undefined;
+  @Prop() public haulsSettings!: BoatnetHaulsSettings;
+  @Prop() public haulsData!: any[];
+  @Prop() public selected: any;
   private searchText = '';
-  private columns = [
-    {
-      name: 'haulNum',
-      required: true,
-      label: 'Haul #',
-      align: 'left',
-      field: 'haulNum',
-      sortable: true
-    },
-    {
-      name: 'weightMethod',
-      align: 'center',
-      label: 'WM',
-      field: (row: any) => row.observerTotalCatch.weightMethod,
-      sortable: true
-    },
-    {
-      name: 'gearPerf',
-      align: 'center',
-      label: 'Gear Perf',
-      field: 'gearPerformance',
-      sortable: true
-    },
-    {
-      name: 'targetStrategy',
-      align: 'center',
-      label: 'Strat',
-      field: 'targetStrategy',
-      sortable: true
-    },
-    {
-      name: 'gearType',
-      align: 'center',
-      label: 'Gear Type',
-      field: 'gearType',
-      sortable: true
-    },
-    {
-      name: 'setDate',
-      align: 'center',
-      label: 'Set Time', // TODO: this needs logic to get first and last
-      field: (row: any) =>
-        moment(row.locations[0].date).format('hh:mm MM/DD/YY'),
-      sortable: true
-    },
-    {
-      name: 'upDate',
-      align: 'center',
-      label: 'Up Time', // TODO: this needs logic to get first and last
-      field: (row: any) =>
-        moment(row.locations[1].date).format('hh:mm MM/DD/YY'),
-      sortable: true
-    },
-    {
-      name: 'otcWeight',
-      align: 'center',
-      label: 'OTC Wt',
-      field: (row: any) => row.observerTotalCatch.value,
-      sortable: true
-    },
-
-    {
-      name: 'errors',
-      align: 'center',
-      label: 'Errors',
-      field: (row: any) => 0, // TODO Error calc
-      sortable: true
-    }
-  ];
-  private data: any = [];
-
-  private exampleHaul: WcgopHaul;
-  private exampleHaul2: WcgopHaul;
-
-  get programName() {
-    // Computed Property
-    return this.programInfo ? this.programInfo.name : '';
-  }
-
-  constructor() {
-    super();
-
-    const locationExample: LocationEvent = {
-      location: {
-        // GeoJSON https://tools.ietf.org/html/rfc7946
-        type: 'Point',
-        coordinates: [124.6, 10.1]
-      },
-      date: moment()
-        .subtract(1, 'days')
-        .format()
-    };
-
-    const locationExample2: LocationEvent = {
-      location: {
-        // GeoJSON https://tools.ietf.org/html/rfc7946
-        type: 'Point',
-        coordinates: [124.6, 10.1]
-      },
-      date: moment().format()
-    };
-
-    this.exampleHaul = {
-      _id: '1',
-      type: WcgopHaulTypeName,
-      createdBy: 'test',
-      createdDate: moment().format(),
-      haulNum: 1,
-      observerTotalCatch: {
-        value: 170,
-        units: 'kg',
-        weightMethod: '21'
-      },
-      locations: [locationExample, locationExample2],
-      gearPerformance: '5',
-      targetStrategy: 'ALBC',
-      gearType: 'Trawl'
-    };
-
-    this.exampleHaul2 = {
-      ...this.exampleHaul,
-      haulNum: 2,
-      targetStrategy: 'BANK'
-    };
-    this.data = [this.exampleHaul, this.exampleHaul2];
-  }
 }
 </script>
